zephyr: cros_kb_raw_ite: Fix a issue with ksoh1 variable type
When KEYBOARD_COLS_MAX>15, the value of KSOH_PIN_MASK can exceed the
range of 8-bit value. This causes the compiler to generate an error for
this item.

BUG=b:391036256
TEST=zmake build telith --clobber
TEST=When KEYBOARD_COLS_MAX>15, the build code will be normal.
LOW_COVERAGE_REASON=b:349964233 No test coverage yet.

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6178925
Code-Coverage: Eric Yilun Lin <yllin@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>

+++ zephyr/drivers/cros_kb_raw/cros_kb_raw_ite.c
@@ -127,7 +127,7 @@ static void kb_raw_ite_drive_column_reg_set_v1(const struct device *dev)
 	 *       rest be configured as GPIO output mode. In this case that we
 	 *       disable the ISR in critical section to avoid race condition.
 	 */
-	inst->KBS_KSOH1 &= ~KSOH_PIN_MASK;
+	inst->KBS_KSOH1 &= (uint8_t)~KSOH_PIN_MASK;
 	/* restore interrupts */
 	irq_unlock(key);
 	/* KSO[17:16] pins output low */
